requirements.txt
openai
pyyaml
tqdm